Add Class Bindings for Correct/Incorrect
Make questions into their own component.
Make each answer to a question its own component.

Add class bindings that change the background color for the answer
button to red or green, depending on if the answer is the correct
response or not.

Add score logic that increments/decrements the score by 5 with a correct
or incorrect response.

Add answered property that determines if the question has already been
answered or not.