Add Question_Answer Model
Adds new question_answer model, which is used to determine whether or
not a user has answered a particular question before.

Adds logic to only give/subtract points for a question if a user has not
previously answered the question before.

Answering a question for the first time creates a POST request, adding a
new entry to the question_responses resource on the back end. If a user
tries to answer a question again, points will not be earned, as it will
check to see if the question has been answered before.